% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_res.R
\name{combine_res}
\alias{combine_res}
\title{Combine results from two separate simulation studies}
\usage{
combine_res(res1, res2)
}
\arguments{
\item{res1}{Simulation object 1}

\item{res2}{Simulation object 2}
}
\description{
Combine results from two separate simulation studies
}
\details{
This can be either a diagaccSim1 or diagaccSim2 object.
}
