% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{estimation}
\alias{estimation}
\title{Main function of the diagacc package}
\usage{
estimation(df, ID, index, reference, cutoff, reader, method, logit_transform)
}
\description{
`estimation` returns the diagnostic accuracy estimates of the package
}
\details{
A typical function call will be of the form
estimation(df, ID = "Patient", index = "test", reference = "reference", logit_transform = TRUE)
where df is the used dataframe. ID is a reference variable for subjects. index is the test for which
accuracy will be estimated. reference is the gold standard against which the index test is compared.
logit_transform is an optional transform of the model to ensure estimates being in the unit interval (recommended).
cutoff is the diagnostic cutoff value which, if specified, will be used to estimate sensitivity and specificity.
reader and method are (not yet functional!) reference variables for factorial designs.
}
