% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_fm.R
\name{fit_fm}
\alias{fit_fm}
\title{Fit finite mixture model}
\usage{
fit_fm(X, n.sample = 2000, n.chains = 1, n.thin = 1,
  n.burnin = 800, n.adapt = 200, raw = FALSE,
  runjags.method = "rjags", silent = FALSE)
}
\arguments{
\item{X}{(matrix) Data set.}

\item{n.sample}{Number of MCMC samples.}

\item{n.chains}{Number of chains.}

\item{n.thin}{Thinning value.}

\item{n.burnin}{Number of burn-in.}

\item{n.adapt}{Number of adaptation samples.}

\item{raw}{(logical) Return the randomLCA or runjags object?}

\item{runjags.method}{Parallel or normal method. See runjags documentation.}

\item{silent}{(logical) Suppress output.}
}
\description{
Fit finite mixture model
}
