% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fitted.diagaccMod}
\alias{fitted.diagaccMod}
\title{Obtain fitted values}
\usage{
\method{fitted}{diagaccMod}(x, a, b, ...)
}
\arguments{
\item{x}{diagaccMod object}

\item{a}{Item 1}

\item{b}{Item 2}

\item{...}{Not used}
}
\value{
Print fitted values
}
\description{
Obtain fitted values
}
