% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{gen_data}
\alias{gen_data}
\alias{gen_lc}
\alias{gen_lcre}
\alias{gen_fm}
\title{Generate item response data sets}
\usage{
gen_lc(
  n = 250,
  tau = 0.08,
  miss.prop = 0.2,
  seed = NULL,
  name.items = TRUE,
  drop.gs = FALSE
)

gen_lcre(
  n = 250,
  tau = 0.08,
  miss.prop = 0.2,
  seed = NULL,
  sigma = c(1.5, 1.5),
  name.items = TRUE,
  drop.gs = FALSE
)

gen_fm(
  n = 250,
  tau = 0.08,
  miss.prop = 0.2,
  seed = NULL,
  eta = c(0.5, 0.2),
  name.items = TRUE,
  drop.gs = FALSE
)
}
\arguments{
\item{n}{Sample size.}

\item{tau}{Prevalence of the disease, i.e. P(\eqn{\delta=1}).}

\item{miss.prop}{Proportion of missing values in the gold standard item.}

\item{seed}{Random seed.}

\item{name.items}{(Logical) Use item names as set in the options?}

\item{drop.gs}{(Logical) Drop the gold standard item?}

\item{sigma}{The standard deviation of the random effects.}

\item{eta}{(Vector of length 2) Probabilities of correctly classifying
diseased and healthy individuals respectively.}
}
\value{
Returns a data frame.
}
\description{
Generate a data set using the simple latent class model (\code{gen_lc}), the
latent class with random effects model (\code{gen_lcre}), or the finite
mixture model (\code{gen_fm}) as the underlying data generating mechanism.
}
\details{
The parameters are set in the package options.
}
\author{
Haziq Jamil, Elena Erosheva
}
