% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{gen_lc}
\alias{gen_lc}
\title{Latent class model data generation}
\usage{
gen_lc(n = 250, tau = 0.08, miss.prop = 0.2, seed = NULL,
  name.items = TRUE)
}
\arguments{
\item{n}{Sample size.}

\item{tau}{Prevalence of the disease, i.e. \eqn{\text{P}(\delta=1)}.}

\item{miss.prop}{Proportion of missing values in the gold standard item.}

\item{seed}{Random seed.}

\item{name.items}{Logical. Use item names as set in the options?}
}
\value{
Data frame.
}
\description{
Generate a data set using the simple latent class model as the underlying
data generating mechanism.
}
\details{
Drops the last item if miss.prop = 1.
}
\examples{
gen_lc(n = 10)
}
