% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generate_data.R
\name{gen_lcre}
\alias{gen_lcre}
\title{Latent class with random effects model data generation}
\usage{
gen_lcre(n = 250, tau = 0.08, miss.prop = 0.2, seed = NULL,
  sigma = c(1.5, 1.5), name.items = TRUE)
}
\arguments{
\item{n}{Sample size.}

\item{tau}{Prevalence of the disease, i.e. \eqn{\text{P}(\delta=1)}.}

\item{miss.prop}{Proportion of missing values in the gold standard item.}

\item{seed}{Random seed.}

\item{sigma}{The standard deviation of the random effects.}

\item{name.items}{Logical. Use item names as set in the options?}
}
\value{
Data frame.
}
\description{
Generate a data set using the latent class with random effects model as the
underlying data generating mechanism.
}
\details{
Drops the last item if miss.prop = 1.
}
\examples{
gen_lcre(n = 10)
}
