% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.diagaccSim2}
\alias{plot.diagaccSim2}
\title{Plot diagaccSim2 study object}
\usage{
\method{plot}{diagaccSim2}(x, sim.key, type = c("est", "se"), ...)
}
\arguments{
\item{x}{diagaccSim2 object}

\item{sim.key}{The study result to display}

\item{type}{Plot type, estimates or standard error.}

\item{...}{Not used.}
}
\description{
Plot diagaccSim2 study object
}
