% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{print.diagaccMod}
\alias{print.diagaccMod}
\title{Print method for diagaccMod}
\usage{
\method{print}{diagaccMod}(x, ...)
}
\arguments{
\item{x}{diagaccMod object}

\item{...}{Not used}
}
\value{
Print diagaccMod object values
}
\description{
Print method for diagaccMod
}
