% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{run_sim}
\alias{run_sim}
\title{Run a simulation}
\usage{
run_sim(B = 4, n = 250, tau = 0.08, miss.prop = 0.2,
  data.gen = c("lc", "lcre", "fm"), lc.method = "MCMC",
  lcre.method = "MCMC", pb, object = NULL)
}
\arguments{
\item{B}{Number of replications}

\item{n}{(numeric) Sample size}

\item{tau}{(numeric, between 0 and 1) Prevalence}

\item{miss.prop}{(numeric, between 0 and 1) Missing proportion}

\item{data.gen}{The data generating mechanism}

\item{lc.method}{(DEPRECATED) The latent class model estimation method}

\item{lcre.method}{(DEPRECATED) The latent class model with random effects
estimation method}

\item{pb}{Not used}

\item{object}{A diagaccSim1 object}
}
\value{
A diagaccSim1 object. It is a list of length four containing the
  following items: \itemize{\item \code{LC}: A list of length \code{B}
  containing the LC model fit using the \code{fit_lc()} function. \item
  \code{LCRE}: A list of length \code{B} containing the LCRE model fit using
  the \code{fit_lcre()} function. \item \code{FM}: A list of length \code{B}
  containing the LC model fit using the \code{fit_fm()} function. \item \code{sim.settings}: A list of length \code{B}
  containing the LC model fit using the \code{fit_lc()} function.}
}
\description{
Repeatedly fit the LC, LCRE and FM model onto a specific scenaria.
}
\details{
The framework for the simulation is as follows: \enumerate{\item Randomly
generate a specific data set according to the \code{data.gen} mechanism and
the settings as per \code{n}, \code{tau}, and \code{miss.prop}. \item Fit a
LC, LCRE and FM model onto the data set. \item Repeat steps 1--2 a total of
\code{B} number of times.}

Note that it is possible to continue on a simulation run by calling the saved
\code{object} in the argument. See the example section.
}
