% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_parallel.R
\name{run_sim_par}
\alias{run_sim_par}
\title{Title}
\usage{
run_sim_par(object = NULL, B = 4, n = 250, tau = 0.08,
  miss.prop = 0.2, data.gen = c("lc", "lcre", "fm"),
  no.cores = parallel::detectCores(), lc.method = c("MCMC", "EM"),
  lcre.method = c("MCMC", "EM"))
}
\arguments{
\item{object}{A diagaccSim1 object}

\item{B}{Number of replications}

\item{n}{(numeric) Sample size}

\item{tau}{(numeric, between 0 and 1) Prevalence}

\item{miss.prop}{(numeric, between 0 and 1) Missing proportion}

\item{data.gen}{The data generating mechanism}

\item{no.cores}{}

\item{lc.method}{(DEPRECATED) The latent class model estimation method}

\item{lcre.method}{(DEPRECATED) The latent class model with random effects
estimation method}
}
\description{
Title
}
