% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study.R
\name{run_study}
\alias{run_study}
\title{Run a study}
\usage{
run_study(object = NULL, B = 4, n = c(250, 1000), tau = c(0.08,
  0.4), miss.prop = c(0.5, 0.8, 1), lc.method = c("EM", "MCMC"),
  lcre.method = c("EM", "MCMC"))
}
\arguments{
\item{object}{diagaccSim2 object}

\item{B}{Number of replications}

\item{n}{Setting for sample size}

\item{tau}{Setting for prevalence}

\item{miss.prop}{Setting for missing proportion}

\item{lc.method}{Method for fitting LC models}

\item{lcre.method}{Method for fitting LCRE models}
}
\description{
Run a study
}
