% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{sim_res}
\alias{sim_res}
\title{Helper function to provide summary of simulation results}
\usage{
sim_res(x, type = c("est", "bias", "sd"), raw = FALSE)
}
\arguments{
\item{x}{A \code{diagaccSim1} object.}

\item{type}{Choose between \code{"est"}, \code{"bias"} or \code{"sd"}.}

\item{raw}{(Logical) Provide the raw matrix of replicates?}
}
\description{
This is an internal helper function to provide summary of simulation results.
}
