% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_lcre.R
\name{fit_lcre}
\alias{fit_lcre}
\title{Fit latent class with random effects model}
\usage{
fit_lcre(X, method = c("MCMC", "EM"), n.sample = 2000, n.chains = 1,
  n.thin = 1, n.burnin = 800, n.adapt = 200, raw = FALSE,
  runjags.method = "rjags", silent = FALSE, quad.points = 21,
  calcSE = TRUE, gold.std = FALSE)
}
\arguments{
\item{X}{(matrix) Data set.}

\item{method}{EM algorithm or MCMC.}

\item{n.sample}{Number of MCMC samples.}

\item{n.chains}{Number of chains.}

\item{n.thin}{Thinning value.}

\item{n.burnin}{Number of burn-in.}

\item{n.adapt}{Number of adaptation samples.}

\item{raw}{(logical) Return the randomLCA or runjags object?}

\item{runjags.method}{Parallel or normal method. See runjags documentation.}

\item{silent}{(logical) Suppress output.}

\item{quad.points}{(numeric, positive) Number of quadrature points for
randomLCA fit. Check randomLCA documentation.}

\item{calcSE}{(logical) Calculate standard error of estimates for randomLCA fit.}
}
\description{
Fit latent class with random effects model
}
