% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.diagaccSim1}
\alias{plot.diagaccSim1}
\title{Plot for diagaccSim1 object}
\usage{
\method{plot}{diagaccSim1}(x, type = c("est", "se"), ...)
}
\arguments{
\item{x}{diagaccSim1 object.}

\item{type}{Plot type, estimates or standard error.}

\item{...}{Not used.}
}
\description{
Plot for diagaccSim1 object
}
