% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{run_sim}
\alias{run_sim}
\title{Title}
\usage{
run_sim(object = NULL, B = 3, n = 250, tau = 0.08,
  miss.prop = 0.2, data.gen = c("lc", "lcre", "fm"), pb,
  lc.method = c("EM", "MCMC"), lcre.method = c("EM", "MCMC"))
}
\arguments{
\item{object}{A diagaccSim1 object}

\item{B}{Number of replications}

\item{n}{(numeric) Sample size}

\item{tau}{(numeric, between 0 and 1) Prevalence}

\item{miss.prop}{(numeric, between 0 and 1) Missing proportion}

\item{data.gen}{The data generating mechanism}

\item{pb}{Not used}

\item{lc.method}{The latent class model estimation method}

\item{lcre.method}{The latent class model with random effects estimation method}
}
\value{
A diagaccSim1 object
}
\description{
Title
}
